提交了一大波代码

+++ tang-gateway/src/main/java/net/ttcxy/tang/gateway/controller/DtsBlogCommentController.java
@@ -1,17 +1,19 @@
 package net.ttcxy.tang.gateway.controller;
 
 import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.util.StrUtil;
 import com.github.pagehelper.PageInfo;
 import io.swagger.annotations.ApiOperation;
 import net.ttcxy.tang.gateway.core.api.ApiException;
 import net.ttcxy.tang.gateway.core.api.ResponseCode;
 import net.ttcxy.tang.gateway.core.api.ResponseResult;
-import net.ttcxy.tang.gateway.entity.dto.DtsCommentDto;
+import net.ttcxy.tang.gateway.entity.dto.DtsBlogCommentDto;
 import net.ttcxy.tang.gateway.entity.model.DtsBlogComment;
 import net.ttcxy.tang.gateway.entity.param.DtsBlogCommentParam;
 import net.ttcxy.tang.gateway.service.CurrentAuthorService;
 import net.ttcxy.tang.gateway.service.DtsCommentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 /**
@@ -19,6 +21,7 @@ import org.springframework.web.bind.annotation.*;
  */
 @RestController
 @RequestMapping("comment")
+@Validated
 public class DtsBlogCommentController {
 
     @Autowired
@@ -27,10 +30,11 @@ public class DtsBlogCommentController {
     @Autowired
     private DtsCommentService dtsCommentService;
 
-    @PostMapping("insert")
+    @PostMapping
     @ApiOperation("添加博客评论")
     public ResponseResult<?> insertComment(@RequestBody DtsBlogCommentParam commentParam){
         DtsBlogComment blogComment = BeanUtil.toBean(commentParam, DtsBlogComment.class);
+        blogComment.setAuthorId(currentAuthorService.getAuthorId());
         int count = dtsCommentService.insertComment(blogComment);
         if (count  > 0){
             return ResponseResult.success();
@@ -38,22 +42,30 @@ public class DtsBlogCommentController {
         throw new ApiException(ResponseCode.FAILED);
     }
 
-    @DeleteMapping("delete")
+    @DeleteMapping("{commentId}")
     @ApiOperation("删除博客评论")
-    public ResponseResult<?> deleteComment(@RequestBody DtsBlogCommentParam commentParam){
-        DtsBlogComment blogComment = BeanUtil.toBean(commentParam, DtsBlogComment.class);
-        String commentId = blogComment.getBlogCommentId();
-        int count = dtsCommentService.deleteComment(commentId);
-        if (count > 0){
-            return ResponseResult.success();
+    public ResponseResult<?> deleteComment(@PathVariable("commentId") String commentId){
+        String authorId = currentAuthorService.getAuthorId();
+        DtsBlogComment blogComment = dtsCommentService.selectCommentById(commentId);
+
+        if (blogComment == null){
+            throw new ApiException(ResponseCode.VALIDATE_FAILED);
         }
+
+        if (StrUtil.equals(blogComment.getAuthorId(),authorId)) {
+            int count = dtsCommentService.deleteComment(commentId);
+            if (count > 0){
+                return ResponseResult.success();
+            }
+        }
+
         throw new ApiException(ResponseCode.FAILED);
     }
 
     @GetMapping("list")
     @ApiOperation("查询博客评论")
-    public ResponseResult<PageInfo<DtsCommentDto>> selectComment(@RequestParam("blogId") String blogId){
-        PageInfo<DtsCommentDto> commentDtoPageInfo = dtsCommentService.selectComments(blogId, 1, 10);
+    public ResponseResult<PageInfo<DtsBlogCommentDto>> selectComment(@RequestParam("blogId") String blogId,@RequestParam("page") Integer page){
+        PageInfo<DtsBlogCommentDto> commentDtoPageInfo = dtsCommentService.selectComments(blogId, page, 10);
         return ResponseResult.success(commentDtoPageInfo);
     }
